+++ Flow/src/utopia/flow/util/console/Command.scala
@@ -10,6 +10,11 @@ object Command
 {
 	// ATTRIBUTES   ---------------------
 	
+	/**
+	 * A regular expression for splitting namespace elements from the command name / input
+	 */
+	val namespaceSplitRegex = Regex.escape(':')
+	
 	private val whiteSpaceOutsideQuotesR = Regex.whiteSpace.ignoringQuotations
 	
 	
@@ -24,7 +29,7 @@ object Command
 	 * @return A new command
 	 */
 	def withoutArguments(name: String, alias: String = "", help: String = "")(execute: => Unit) =
-		new Command(name, alias, help = help)(_ => execute)
+		_apply(name, alias, CommandArgumentsSchema.empty, help) { _ => execute }
 	/**
 	 * Creates a new command utilizing an existing command arguments schema
 	 * @param schema Arguments schema
@@ -35,29 +40,41 @@ object Command
 	 * @return A new command
 	 */
 	def withSchema(schema: CommandArgumentsSchema, name: String, alias: String = "", help: String = "")
-	              (execute: CommandArguments => Unit) = new Command(name, alias, schema, help)(execute)
+	              (execute: CommandArguments => Unit) =
+		_apply(name, alias, schema, help)(execute)
 	
 	/**
 	 * Creates a new command
-	 * @param name Name of this command
+	 * @param name Name of this command. May include a namespace: e.g. "namespace1:namespace2:command-name"
 	 * @param alias Alias for this command (default = empty = no alias)
 	 * @param help A helpful description of this command (default = empty = no description)
 	 * @param args Schemas for the arguments accepted by this command
 	 * @param execute Function to be called when this command should be executed. Accepts processed command arguments.
 	 * @return A new command
 	 */
-	def apply(name: String, alias: String = "", help: String = "")
-	         (args: ArgumentSchema*)
+	def apply(name: String, alias: String = "", help: String = "")(args: ArgumentSchema*)
 	         (execute: CommandArguments => Unit) =
-		new Command(name, alias, CommandArgumentsSchema(args), help)(execute)
+		_apply(name, alias, CommandArgumentsSchema(args), help)(execute)
+	
+	private def _apply(name: String, alias: String, schema: CommandArgumentsSchema, help: String)
+	                  (execute: CommandArguments => Unit) =
+	{
+		val nameParts = namespaceSplitRegex.split(name)
+		new Command(nameParts.last, alias, nameParts.dropRight(1), schema, help)(execute)
+	}
 }
 
 /**
  * Commands are used for interacting with users in a console. Usually commands accept arguments of some kind.
+ * @param name Name of this command
+ * @param alias Alias for this command. May be empty.
+ * @param namespace Namespace for discerning between commands with the same name. Default = empty.
+ * @param argumentsSchema Command argument definitions
+ * @param help A description of what this command does. May be empty.
  * @author Mikko Hilpinen
  * @since 10.10.2021, v1.12.2
  */
-class Command(val name: String, val alias: String = "",
+class Command(val name: String, val alias: String = "", val namespace: Seq[String] = Empty,
               val argumentsSchema: CommandArgumentsSchema = CommandArgumentsSchema.empty, val help: String = "")
              (execute: CommandArguments => Unit)
 	extends ArgumentMatchable